test: 정적 리소스 컨트롤러 분리에 따른 테스트 이동 및 추가
