+++ tomcat/src/test/java/com/techcourse/controller/FrontControllerTest.java
@@ -24,36 +24,6 @@ class FrontControllerTest {
         frontController = FrontController.getInstance();
     }
 
-    @DisplayName("올바른 리소스에 대해 200 응답을 반환한다.")
-    @Test
-    void findResource() throws Exception {
-        // given
-        RequestLine requestLine = RequestLine.from("GET /css/styles.css HTTP/1.1");
-        HttpHeaders headers = HttpHeaders.from(List.of(
-                "Host: localhost:8080 ",
-                "Connection: keep-alive "
-        ));
-        HttpRequest httpRequest = new HttpRequest(requestLine, headers, new RequestBody());
-        HttpResponse httpResponse = new HttpResponse();
-
-        // when
-        frontController.service(httpRequest, httpResponse);
-
-        // then
-        final URL resource = getClass().getClassLoader().getResource("static/css/styles.css");
-        String expectedResponseLine = "HTTP/1.1 200 OK \r\n";
-        String expectedContentLength = "Content-Length: 211991 \r\n";
-        String expectedContentType = "Content-Type: text/css \r\n";
-        String expectedResponseBody = new String(Files.readAllBytes(new File(resource.getFile()).toPath()));
-
-        assertThat(httpResponse.serialize()).contains(
-                expectedResponseLine,
-                expectedContentLength,
-                expectedContentType,
-                expectedResponseBody
-        );
-    }
-
     @DisplayName("존재하지 않는 리소스에 접근하면 404 응답을 반환한다.")
     @Test
     void wrongURL() throws Exception {
